requirements.txt
django-argparse-command
requests
beautifulsoup4
chartkick
httplib2
